move iter_control into regression package, remove redundant r0 from TRME_RR
